Add password protection for the slots (#2)
# Description

With this PR, Ghoti can be configured with Users (user/password) in
order to protect slots individually for reading and writing.

The configuration file can have a section users, with the username and
password:

```
users:
  pepe: password
  sam: otherPassword
```

And then you can configure the slots using these users:

```
slot000:
  kind: simple_memory
  users:
    pepe: r
    sam: w
```

There are three configurations:
- r for read only
- w for write only
- a for all access

The configurations are per slot. In order to login two commands need to
be sent, the `u` command to define the username and then the `p` command
to define the password.

```
> upepe
> vpepe
> ppassword
> vpepe
```
 
With a response having the `v` value response type followed by the
username, it is confirmed that the user is logged in. In order to use
this feature I recommend to use TLS to encrypt the communication as the
password will be travelling without any encoding.

# Testing
Added coverage for the new feature.