Add ticker slot (#30)
This pull request introduces a new `ticker` slot type to the
application. The changes include updates to the documentation, the
addition of the ticker slot implementation, and corresponding tests.

### Documentation Updates:
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R161-R168):
Added configuration examples for the `leaky_bucket` and `ticker` slots.
Updated the description and configuration details for the `ticker` slot.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R161-R168)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L208-R235)

### Implementation of Ticker Slot:
*
[`internal/slots/slots.go`](diffhunk://#diff-2ac0d3ea5b0fcb5f7e98ab6313bc8f69f8806ce61bf1a682ec55449a5a103e11R94-R120):
Added logic to handle the creation of a `ticker` slot, including
validation for `initial_value` and `refresh_rate`.
*
[`internal/slots/ticker.go`](diffhunk://#diff-be43318241966c6f8216039115f41abaebceecf525f7aeffa8fd91f13bb4fab6R1-R74):
Implemented the `tickerSlot` struct and its methods, including `Read`,
`CanRead`, `CanWrite`, and `Write`.

### Testing:
*
[`internal/slots/ticker_test.go`](diffhunk://#diff-c6be442c7c4ee7523264acca098ec52bf7e1ab3622026e9a9e5d60d0a87d0759R1-R118):
Added tests for the `ticker` slot, covering initialization, read/write
permissions, ticking behavior, and handling of missing configuration.