Add async goroutine to handle writes (#31)
# Description
This PR adds a goroutine and channels to handle the writes to every
connection.

Basicall, now every connection has an event channel where messages to
the clients can be sent. This channel is a buffer channel, this allows
us to handle the writes up to a point, because if the queue is too long
it means that the client is not consuming the messages fast enough.

This also allows to add a write timeout per message sent (this will be
configurable in a future PR).

The event has certain fields that allow it to be dismissed if it has
been in the queue for too long. Once the message was sent (or failed
sending), there is also a callback channel that can be used to notify
about the status of this message. The events also have an internal ID
that can be used to identify which message has failed or succeed.