License update to Apache License  V2.0 (#48)
This pull request updates the licensing terms for the project by
replacing the existing custom license with the Apache License 2.0. The
most important changes include the adoption of a standardized
open-source license and the removal of the previous custom license.

### Licensing updates:

* Added the Apache License 2.0 to the project, replacing the previous
custom license. This change standardizes the licensing terms and
provides a widely recognized open-source license. (`LICENSE`)

* Removed the previous custom license file (`LICENSE.md`) that outlined
unique terms and conditions specific to the project. This simplifies the
licensing approach.