Add release workflows (#50)
This pull request introduces significant improvements to the CI/CD
pipeline and build process for the `ghoti` project. The changes include
new GitHub Actions workflows for release automation and package updates,
as well as enhancements to the `Makefile` for multi-platform builds and
release packaging.

### CI/CD Enhancements

* Added a new GitHub Actions workflow (`.github/workflows/release.yml`)
to automate the release process when a new tag is pushed. This workflow
builds the project, creates release packages, and publishes them to
GitHub Releases with detailed release notes.
* Introduced another GitHub Actions workflow
(`.github/workflows/go-pkg-update.yml`) to request updates on
`pkg.go.dev` after a successful release. This ensures that the latest
version is indexed and available for Go developers.

### Build Process Improvements

* Enhanced the `Makefile` with a new `build-all` target to build the
`ghoti` binary for multiple platforms (Linux, macOS, and Windows) and
architectures (amd64, arm64, 386).
* Added a `release` target to the `Makefile` for creating release
packages. This includes zipping binaries along with the `README.md` file
for each platform and architecture.