Replace copy of http.Server with reference (#55)
Presently on master, `joinServer.Stop()` does not stop the HTTP server.
I proved this by adding a `sync.WaitGroup` to `joinServer` which caused
`Stop()` to stall because `Shutdown` was being called on a different
copy of the server from the copy on which Start was called.

I found this because my editor highlighted an instance of a copied lock
and when this happens it is common to find other issues.


![image](https://github.com/user-attachments/assets/344b1d42-6ed1-46a7-b8e1-7a6f97f9ed40)

The solution is to make the `server` property a `*http.Server` so there
is only one.